Type Definition polkadot_runtime::AllPalletsWithSystemReversed

pub type AllPalletsWithSystemReversed = (MessageQueue, XcmPallet, Crowdloan, Auctions, Slots, Registrar, ParaAssignmentProvider, ParasSlashing, ParasDisputes, ParaSessionInfo, Hrmp, Dmp, Initializer, Paras, ParaScheduler, ParaInherent, ParaInclusion, ParasShared, Configuration, ParachainsOrigin, FastUnstake, NominationPools, VoterList, ElectionProviderMultiPhase, ChildBounties, Bounties, Multisig, Proxy, Identity, Utility, Vesting, Claims, Whitelist, Origins, Referenda, ConvictionVoting, Treasury, AuthorityDiscovery, ImOnline, Grandpa, Session, Historical, Offences, Staking, Authorship, TransactionPayment, Balances, Indices, Timestamp, Babe, Preimage, Scheduler, System);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletsWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order.