Type Definition polkadot_runtime::AllPalletsWithoutSystemReversed

pub type AllPalletsWithoutSystemReversed = (XcmPallet, Crowdloan, Auctions, Slots, Registrar, ParasDisputes, ParaSessionInfo, Hrmp, Ump, Dmp, Initializer, Paras, ParaScheduler, ParaInherent, ParaInclusion, ParasShared, Configuration, ParachainsOrigin, NominationPools, VoterList, ElectionProviderMultiPhase, Tips, ChildBounties, Bounties, Multisig, Proxy, Identity, Utility, Vesting, Claims, Treasury, TechnicalMembership, PhragmenElection, TechnicalCommittee, Council, Democracy, AuthorityDiscovery, ImOnline, Grandpa, Session, Historical, Offences, Staking, Authorship, TransactionPayment, Balances, Indices, Timestamp, Babe, Preimage, Scheduler);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.