Type Definition polkadot_runtime::xcm_config::LocalAssetTransactor

pub type LocalAssetTransactor = XcmCurrencyAdapter<Balances, IsConcrete<DotLocation>, SovereignAccountOf, AccountId, CheckAccount>;

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like MultiLocation and MultiAsset.

Ours is only aware of the Balances pallet, which is mapped to DotLocation.