Crate polkadot_runtime

The Polkadot runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

Constant values used within the runtime.

Structs

AdvertisedXcmVersion

Ancestry

Our XCM location ancestry - i.e. what, if anything, Parent means evaluated in our context. Since Polkadot is a top-level relay-chain, there is no ancestry.

AnnouncementDepositBase

AnnouncementDepositFactor

BagThresholds

BaseFilter

BaseXcmWeight

The amount of weight an XCM operation takes. This is a safe overestimate.

BasicDeposit

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

CandidacyBond

CheckAccount

The check account, which holds any native assets that have been teleported out and not back in (yet).

CooloffPeriod

CouncilBodyId

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CrowdloanId

DataDepositPerByte

DepositBase

DepositFactor

DesiredMembers

13 members initially, to be increased to 23 eventually.

DesiredRunnersUp

DotLocation

The location of the DOT token, from the context of this chain. Since this token is native to this chain, we make it synonymous with it and thus it is the Here location, which means “equivalent to the context”.

EnactmentPeriod

EndingPeriod

EpochDuration

ExistentialDeposit

ExpectedBlockTime

FastTrackVotingPeriod

FieldDeposit

FirstMessageFactorPercent

GenesisConfig

ImOnlineUnsignedPriority

IndexDeposit

InstantAllowed

LaunchPeriod

LeaseOffset

LeasePeriod

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxInstructions

Maximum number of instructions in a single XCM fragment. A sanity check against weight calculations getting too crazy.

MaxKeys

MaxLocks

MaxMemoLength

MaxNominatorRewardedPerValidator

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxPending

MaxProposals

MaxProxies

MaxRegistrars

MaxReserves

MaxRetries

MaxScheduledPerBlock

MaxSignatories

MaxSubAccounts

MaxVotes

MaximumReasonLength

MaximumSchedulerWeight

MinContribution

MinVestedTransfer

MinimumDeposit

MinimumPeriod

NposCompactSolution16

NposSolutionPriority

OffchainRepeat

OffendingValidatorsThreshold

OnlyParachains

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

Origin

The runtime origin type representing the origin of a call.

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDataByteDeposit

ParaDeposit

ParathreadDeposit

PhragmenElectionPalletId

Polkadot

PolkadotForStatemint

PolkadotNetwork

The Polkadot network ID. This is named.

Prefix

PreimageByteDeposit

ProposalBond

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

RemoveKeysLimit

ReportLongevity

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

SS58Prefix

SampleLength

SessionKeys

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SolutionImprovementThreshold

SpendPeriod

StakingBagsListMigrationV8

SubAccountDeposit

SubmissionDeposit

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TermDuration

Weekly council elections; scaling up to monthly eventually.

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

TreasuryPalletId

UncleGenerations

UnsignedPhase

Version

VoterSnapshotPerBlock

Whilst UseNominatorsAndUpdateBagsList or UseNominatorsMap is in use, this can still be a very large value. Once the BagsList is in full motion, staking might open its door to many more nominators, and this value should instead be what is a “safe” number (e.g. 22500).

VotingBondBase

VotingBondFactor

VotingPeriod

XcmConfig

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

EPMCall

Contains one variant per dispatchable that can be called by an extrinsic.

Event

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

StakerStatus

Indicates the initial status of the staker.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

VERSION

Runtime version (Polkadot).

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

Type Definitions

Address

The address format for describing accounts.

AllModulesDeprecated

All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystemDeprecated

All modules included in the runtime as a nested tuple of types.

AllPallets

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

BagsList

Balances

BalancesConfig

Barrier

The barriers one of which must be passed for an XCM message to be executed.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bounties

Claims

ClaimsConfig

Configuration

ConfigurationConfig

Council

CouncilCollective

CouncilConfig

Crowdloan

Democracy

DemocracyConfig

Dmp

ElectionProviderMultiPhase

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Hrmp

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Initializer

LocalAssetTransactor

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like MultiLocation and MultiAsset.

LocalOriginToLocation

Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

Multisig

Offences

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Paras

ParasConfig

ParasShared

PhragmenElection

PhragmenElectionConfig

Proxy

Registrar

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Slots

SovereignAccountOf

The canonical means of converting a MultiLocation into an AccountId, used when we want to determine the sovereign account controlled by a location.

Staking

StakingConfig

System

SystemConfig

TechnicalCollective

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tips

TransactionPayment

Treasury

TreasuryConfig

TrustedTeleporters

Ump

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig

XcmPallet

XcmPalletConfig

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.