Crate polkadot_runtime

The Polkadot runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

governance

New governance configurations for the Polkadot runtime.

migrations

The runtime migrations per release.

xcm_config

XCM configuration for Polkadot.

Macros

impl_elections_weights

Implements the weight types for the elections module and a specific runtime. This macro should not be called directly; use [impl_runtime_weights] instead.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

BagThresholds

BasicDeposit

BetterUnsignedThreshold

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockWeights

Block weights base values and limits.

BondingDuration

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

ChildBountyValueMinimum

CouncilSpendOriginMaxAmount

CrowdloanId

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

DataDepositPerByte

DepositBase

DepositFactor

ElectionBounds

We take the top 22500 nominators as electing voters and all of the validators as electable targets. Whilst this is the case, we cannot and shall not increase the size of the validator intentions.

EndingPeriod

EpochDuration

EraPayout

ExistentialDeposit

ExpectedBlockTime

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

FieldDeposit

ImOnlineUnsignedPriority

IndexDeposit

InitiateNominationPools

LeaseOffset

LeasePeriod

MaxActiveChildBountyCount

MaxActiveValidators

Setup election pallet to support maximum winners upto 1200. This will mean Staking Pallet cannot have active validators higher than this count.

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxElectingVoters

MaxKeys

MaxLocks

MaxMemoLength

MaxNominations

MaxNominatorRewardedPerValidator

MaxPeerInHeartbeats

MaxPending

MaxPointsToBalance

MaxProxies

MaxRegistrars

MaxReserves

MaxRetries

MaxScheduledPerBlock

MaxSetIdSessionEntries

MaxSignatories

MaxSubAccounts

MaximumReasonLength

MaximumSchedulerWeight

MessageProcessor

Message processor to handle any messages that were enqueued into the MessageQueue pallet.

MessageQueueHeapSize

MessageQueueMaxStale

MessageQueueServiceWeight

Amount of weight that can be spent per block to service messages.

MinContribution

MinVestedTransfer

MinimumPeriod

NoPreimagePostponement

NominationPoolsMigrationV4OldPallet

NposCompactSolution16

NposSolutionPriority

OffchainRepeat

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution submission.

OffendingValidatorsThreshold

OnChainSeqPhragmen

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDataByteDeposit

ParaDeposit

ParasUnsignedPriority

ParathreadDeposit

ParityDbWeight

ParityDB can be enabled with a feature flag, but is still experimental. These weights are available for brave runtime engineers who may want to try this out as default.

PoolsPalletId

Prefix

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMaximum

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

RemoveKeysLimit

ReportLongevity

RewardCurve

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

RootSpendOriginMaxAmount

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SampleLength

SessionKeys

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedMaxRefunds

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SpendPeriod

SubAccountDeposit

SubmissionDeposit

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

TreasuryPalletId

UnsignedPhase

UnvestedFundsAllowedWithdrawReasons

Version

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

EPMCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

StakerStatus

Representation of the status of a staker.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 1% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 1%.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block time. The storage proof size is not limited so far.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

VERSION

Runtime version (Polkadot).

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bounties

ChildBounties

Claims

ClaimsConfig

Configuration

ConfigurationConfig

ConvictionVoting

Crowdloan

Dmp

ElectionProviderMultiPhase

Executive

Executive: handles dispatch to the various modules.

FastUnstake

GenesisConfigDeprecated

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Hrmp

HrmpConfig

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Initializer

MessageQueue

Migrations

All migrations that will run on the next runtime upgrade.

Multisig

NominationPools

NominationPoolsConfig

Offences

Origins

ParaAssignmentProvider

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Paras

ParasConfig

ParasDisputes

ParasShared

ParasSlashing

Preimage

Proxy

Referenda

Registrar

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Slots

Staking

StakingConfig

System

SystemConfig

Timestamp

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig

VoterList

Whitelist

XcmPallet

XcmPalletConfig