Crate polkadot_runtime

The Polkadot runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants
Constant values used within the runtime.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

BaseFilter

BasicDeposit

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

Burn

CandidacyBond

CooloffPeriod

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DataDepositPerByte

DepositBase

DepositFactor

DesiredMembers
13 members initially, to be increased to 23 eventually.

DesiredRunnersUp

DisabledValidatorsThreshold

EnactmentPeriod

EpochDuration

ExistentialDeposit

ExpectedBlockTime

Fallback

FastTrackVotingPeriod

FieldDeposit

GenesisConfig

ImOnlineUnsignedPriority

IndexDeposit

InstantAllowed

LaunchPeriod

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxLocks

MaxNominatorRewardedPerValidator

MaxPending

MaxProposals

MaxProxies

MaxRegistrars

MaxReserves

MaxRetries

MaxScheduledPerBlock

MaxSignatories

MaxSubAccounts

MaxVotes

MaximumReasonLength

MaximumSchedulerWeight

MinVestedTransfer

MinimumDeposit

MinimumPeriod

NposCompactSolution16

NposSolutionPriority

OffchainRepeat

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParathreadDeposit

PhragmenElectionPalletId

Prefix

PreimageByteDeposit

ProposalBond

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

ReportLongevity

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SS58Prefix

SessionKeys

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SolutionImprovementThreshold

SpendPeriod

SubAccountDeposit

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMembershipStoragePrefixMigration
Migrate from Instance1Membership to the new pallet prefix TechnicalMembership

TechnicalMotionDuration

TermDuration
Weekly council elections; scaling up to monthly eventually.

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

TreasuryPalletId

UncleGenerations

UnsignedPhase

Version

VotingBondBase

VotingBondFactor

VotingPeriod

Enums

BalancesCall
Contains one variant per dispatchable that can be called by an extrinsic.

Call

EPMCall
Contains one variant per dispatchable that can be called by an extrinsic.

Event

OriginCaller

ProxyType
The type used to represent the kinds of proxying allowed.

StakerStatus
Indicates the initial status of the staker.

TimestampCall
Contains one variant per dispatchable that can be called by an extrinsic.

Constants

BABE_GENESIS_EPOCH_CONFIG
The BABE epoch configuration at genesis.

VERSION
Runtime version (Polkadot).

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage
Complex storage builder stuff.

Functions

native_version
Native version.

Type Definitions

Address
The address format for describing accounts.

AllModules	Deprecated
All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem	Deprecated
All modules included in the runtime as a nested tuple of types.

AllPallets
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem
All pallets included in the runtime as a nested tuple of types.

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

Bounties

Claims

ClaimsConfig

Council

CouncilCollective

CouncilConfig

Democracy

DemocracyConfig

ElectionProviderMultiPhase

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Multisig

Offences

PhragmenElection

PhragmenElectionConfig

Proxy

Scheduler

Session

SessionConfig

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Staking

StakingConfig

System

SystemConfig

TechnicalCollective

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tips

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig